# Add option to run compiler tests in parallel