Given the `--parallel` option, the compiler test suite can now spawn one process per logical CPU core and split the tests among them. The goal was to make this 100% interchangeable with a normal run, producing the same output except test ordering. Reduces the time of `npm run test:compiler` from 32 to <del>17</del> 13 seconds on my machine using <del>16</del> 7 workers.