@MaxGraey 2019-09-27
I wonder how this different with `worker_threads` in performance? I think `cluster`/`child process fork` have bigger overhead