@dcodeIO 2019-09-27
From what I can tell so far this can be speed up by reducing the number of workers from 16 to 8, matching the number of physical cores, at least on my setup (from 17 to 14 seconds) and by running the std/math and std/typedarray tests, which take the most time, first (from 14 to 12 seconds).