@MaxGraey 2019-09-27
Try also numWorkers = prisicalCores - 1.