@dcodeIO 2019-09-27
Ah, right. That's 0.5 seconds faster than half the physical cores :)