@dcodeIO 2019-09-27
Last commit brings it down to 13 seconds without reordering tests.