@dcodeIO 2019-09-27
From a node profile and a CPU profile I learned that most time is spent in either I/O, compiling a JS file, running the compiler or inside of Binaryen, while all cores spike at 100% at first with some dropping earlier than others. That's most likely because the remaining few long-running tests need to finish while some cores are already idle. I also found that running a non-parallel test utilizes one core at 100% and four at about 40%, so some stuff inside of node/V8 is threaded even in this case, even though the process presents itself as single-threaded. Explains while it's not a NUM_CORES times speedup.