@MaxGraey 2019-09-27
What if replace [this](https://github.com/AssemblyScript/assemblyscript/blob/ecfc957ea8fa75d3bb68d40450b7e324f31ba652/tests/compiler.js#L317) to `instantiateStreaming`?