@dcodeIO 2019-09-28
Does that work in node? If you want, feel free to improve this even more :)