@MaxGraey 2019-09-28
wait) I'm totally forgot that node.js does't support `instantiateStreaming`=)