@MaxGraey 2019-09-28
But we still have async `WebAssembly.instantiate`. I checked this PR on my machine and all my cpus loads only on `~60-70%` so we still have gap for improvements. I guess it should be I/O.