@dcodeIO 2019-09-29
Suggesting to merge this as-is so everyone can try it on their machines, potentially submitting additional improvements. It's not enabled on CI yet, so that should be fine. Wdyt?